fix: Refactored our api directory structure using express router. The /signup route has had a small fix with missing params as well.
