+++ application/client/src/pages/SignUp/SignUpForm.js
@@ -38,10 +38,12 @@ const SignUpForm = () => {
 
     try {
       console.log('Sending registration request...')
-      const response = await Axios.post('https://workwaves-jm2b5.ondigitalocean.app/api/registration', {
-        username: user,
-        password: pwd,
-        email: email
+      const response = await Axios.post('https://workwaves-jm2b5.ondigitalocean.app/api/registration', null, {
+        params: {
+          username: user,
+          password: pwd,
+          email: email
+        }
       })
 
       if (response.status === 201) {
